+++ src/cmd/ai_module/mod.rs
@@ -1,4 +1,3 @@
 pub mod image;
 pub mod transcript;
 pub mod translation;
-pub mod translation_embed;
