+++ src/cmd/lang_struct/embed/error.rs
@@ -11,5 +11,6 @@ pub struct ErrorLocalisedText {
     pub error_request: String,
     pub error_no_avatar: String,
     pub error_parsing_json: String,
+    pub error_url: String,
 }
